FIN-34 refactoring of DashboardChart component (#14)
